Replaced app.js with one that includes openapi-based endpoints
